requirements.txt
typer>=0.12
rich>=13.7
pydantic>=2.7
SQLAlchemy>=2.0
tabulate>=0.9
pytest>=8.0
pytest-cov>=5.0
PySide
